pyproject.toml
[tool.poetry]
name = "manhwa-checker"
version = "0.1.0"
description = ""
authors = ["Alper <itsc0508@gmail.com>"]
readme = "README.md"
packages = [{include = "manhwa_checker"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = {extras = ["async"], version = "^2.2.2"}
beautifulsoup4 = "^4.11.1"
flask-sqlalchemy = "^3.0.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pytest-asyncio = "^0.19.0"
aiohttp = "^3.8.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
